BaytesInference

Documentation for BaytesInference.

BaytesInference.BaytesInference

BaytesInference.check_pf

BaytesInference.compute_escore

BaytesInference.compute_ℓscore

BaytesInference.flatten_chains

BaytesInference.flatten_index

BaytesInference.get_vals

BaytesInference.hpd_grid

BaytesInference.plotChain

BaytesInference.plotDiagnostics

BaytesInference.plotLatent

BaytesInference.plotPosteriorPrediction

BaytesInference.plot_escore

BaytesInference.plot_logscore

BaytesInference.BaytesInference — Module

Bayesian inference on state space models

BaytesInference.check_pf — Method

check_pf(_rng, pf, objective, θ; trials, exact, dmax, plotsize, param_color)

Check log target estimate variance for a interval of different model parameter values.

Examples

BaytesInference.compute_escore — Method

compute_escore(prediction, dataₜ)

Compute escore of trace given data.

Examples

BaytesInference.compute_ℓscore — Method

compute_ℓscore(predictionᵛ, ℓweightsₙ, data; burnin)

Compute log score of trace given data.

Examples

BaytesInference.flatten_chains — Method

Same as above, but only flatten parameter that have valid constraint.

BaytesInference.flatten_index — Method

Flatten subset of model.val for each chain at index 'index'. Useful if indices at different chains are compared.

BaytesInference.get_vals — Method

Return all parameter values, removing first burnin draws.

BaytesInference.hpd_grid — Function

hpd_grid(samples)
hpd_grid(samples, alpha; iter, roundto, ϵ)

Compute HPD interval of a vector of MCMC samples.

Examples

BaytesInference.plotChain — Method

plotChain(trace, algorithm)

Plot samples of trace.

Examples

BaytesInference.plotDiagnostics — Method

plotDiagnostics(diagnostics)

Plot diagnostics of trace.

Examples

BaytesInference.plotLatent — Method

plotLatent(trace, algorithm)

Plot latent state trajectory of trace. Only supported for state space models with HMM or HSMM transitions.

Examples

BaytesInference.plotPosteriorPrediction — Method

plotPosteriorPrediction(trace, algorithm)

Plot posterior predictive samples of trace. Only supported for some algorithms.

Examples

BaytesInference.plot_escore — Method

Plot escore of trace given data.

Examples

BaytesInference.plot_logscore — Method

plot_logscore(trace, smc, data; burnin, delete∞, plotsize, param_color, fontsize, axissize)

Plot log score of trace given data.

Examples