BaytesInference

Documentation for BaytesInference.

BaytesInference.BaytesInference

BaytesInference._compute_escore

BaytesInference.check_pf

BaytesInference.compute_cumℓincrement

BaytesInference.compute_dic

BaytesInference.compute_waic

BaytesInference.compute_ℓscore

BaytesInference.filter_forward

BaytesInference.flatten_chains

BaytesInference.flatten_index

BaytesInference.get_vals

BaytesInference.hpd_grid

BaytesInference.plotChain

BaytesInference.plotChains

BaytesInference.plotDiagnostics

BaytesInference.plotLatent

BaytesInference.plotPosteriorPrediction

BaytesInference.plot_cumℓincrement

BaytesInference.plot_escore

BaytesInference.plot_logscore

BaytesInference.BaytesInference — Module

Bayesian inference on state space models

BaytesInference._compute_escore — Method

_compute_escore(prediction, dataₜ)

Compute escore of trace given data.

Examples

BaytesInference.check_pf — Method

check_pf(_rng, pf, objective, θ)

Check log target estimate variance for a interval of different model parameter values.

Examples

BaytesInference.compute_cumℓincrement — Method

compute_cumℓincrement(diagnostics, effective_iter)

Compute cumulative incremental likelihood from SMC Diagnostics given range 'effective_iter'.

Examples

BaytesInference.compute_dic — Method

compute_dic(ℓlikθ, ℓlikᵥ)
compute_dic(ℓlikθ, ℓlikᵥ, p_dic)

Compute DIC. 'ℓlikθ' is the log target evaluation given posterior mean.'ℓlikᵥ' is a vector of target evaluations p(data | θ_n) for n=1:N. Note that target in this case is likelihood, not posterior.

Examples

BaytesInference.compute_waic — Method

compute_waic(incremental_ℓlikᵥ)
compute_waic(incremental_ℓlikᵥ, t₀)

Compute WAIC. 'incremental_ℓlikᵥ' is a vector of vector, where the inner vector contains the incremental likelihood at each data point, and the outer vector has length equal to the number of MCMC iterations.

Examples

BaytesInference.compute_ℓscore — Method

compute_ℓscore(predictionᵛ, ℓweightsₙ, data)

Compute log score of trace given data.

Examples

BaytesInference.filter_forward — Method

filter_forward(objective)

Check log target estimate variance for a interval of different model parameter values.

Examples

BaytesInference.flatten_chains — Method

Same as above, but only flatten parameter that have valid constraint.

BaytesInference.flatten_index — Method

Flatten subset of model.val for each chain at index 'index'. Useful if indices at different chains are compared.

BaytesInference.get_vals — Method

Return all parameter values, removing first burnin draws.

BaytesInference.hpd_grid — Function

hpd_grid(samples)
hpd_grid(samples, alpha)

Compute HPD interval of a vector of MCMC samples.

Examples

BaytesInference.plotChain — Method

plotChain(trace, algorithm)

Plot samples of trace.

Examples

BaytesInference.plotChains — Method

plotChains(trace, transform)

Plot samples of trace on individual subplots.

Examples

BaytesInference.plotDiagnostics — Method

plotDiagnostics(diagnostics)

Plot diagnostics of trace.

Examples

BaytesInference.plotLatent — Method

plotLatent(trace, algorithm)

Plot latent state trajectory of trace. Only supported for state space models with HMM or HSMM transitions.

Examples

BaytesInference.plotPosteriorPrediction — Method

plotPosteriorPrediction(trace, algorithm)

Plot posterior predictive samples of trace. Only supported for some algorithms.

Examples

BaytesInference.plot_cumℓincrement — Method

plot_cumℓincrement(cumℓincrement)
plot_cumℓincrement(cumℓincrement, dates)
plot_cumℓincrement(cumℓincrement, dates, modelnames)
plot_cumℓincrement(
    cumℓincrement,
    dates,
    modelnames,
    benchmarkmodel
)

Plot cumulative incremental likelihood against each other.

Examples

BaytesInference.plot_escore — Method

Plot escore of trace given data.

Examples

BaytesInference.plot_logscore — Method

plot_logscore(trace, smc, data)

Plot log score of trace given data.

Examples